v1.1.1
特殊代码尝试数据库崩坏时数据同步

+++ largescale/src/main/java/com/waxjx/largescale/service/TeachersService.java
@@ -1,7 +1,9 @@
 package com.waxjx.largescale.service;
 
+import com.waxjx.largescale.Config.ReadWriteRoutingDataSource;
 import com.waxjx.largescale.dao.TeachersMapper;
 import com.waxjx.largescale.model.Teachers;
+import com.waxjx.largescale.util.JdbcUtil;
 import com.waxjx.largescale.util.TeacherSyncUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,6 +24,10 @@ public class TeachersService {
     private Map<String, DataSource> dataSourceMap;
     @Autowired
     private String masterIp;
+    @Autowired
+    private DataSource dataSource;
+
+
 
     /**
      * 根据教师id 查询教师信息 用于教师登录
@@ -38,20 +44,52 @@ public class TeachersService {
 
     @Transactional
     public int updateTeachers(Teachers teachers) {
-        int result = teachersMapper.updateByPrimaryKeySelective(teachers);
-        // update 信息同步到其他数据库
-        TeacherSyncUtil.TeacherSyncUtilUpdate(teachers, dataSourceMap, masterIp);
+        try {
+            String masterUrl = dataSource.getConnection().getMetaData().getURL();
+            masterUrl = JdbcUtil.extractHostFromJdbcUrl(masterUrl);
+            if (masterUrl.equals(masterIp)) {
+                int result = teachersMapper.updateByPrimaryKeySelective(teachers);
 
-        return result;
+                // update 信息同步到其他数据库
+                TeacherSyncUtil.TeacherSyncUtilUpdate(teachers, dataSourceMap, masterIp, masterUrl);
+
+                return result;
+            }else {
+                // update 信息同步到其他数据库
+                TeacherSyncUtil.TeacherSyncUtilUpdate(teachers, dataSourceMap, masterIp, masterUrl);
+                return 1;
+            }
+        }catch (Exception e) {
+            return 0;
+        }
+
     }
 
     @Transactional
     public int insertTeachers(Teachers teachers) {
         try {
-            int result = teachersMapper.insertSelective(teachers);
-            // insert 信息同步到其他数据库
-            TeacherSyncUtil.TeacherSyncUtilInsert(teachers, dataSourceMap, masterIp);
-            return result;
+            // System.err.println(dataSource + "00000000000000000000000000000000000000000000000000000");
+            // System.err.println(dataSource.getConnection().getMetaData().getURL() + "1111111111111111111111111111111111111");
+            // String masterUrl =  "";
+            String masterUrl = dataSource.getConnection().getMetaData().getURL();
+            masterUrl = JdbcUtil.extractHostFromJdbcUrl(masterUrl);
+            // System.err.println(masterUrl+"     222222222222222222222222222222222222222222");
+                    // jdbc:mysql://101.37.38.162:3306/largescale
+            // System.err.println(dataSource);
+            if (masterUrl.equals(masterIp)) {
+                int result = teachersMapper.insertSelective(teachers);
+                System.err.println("第一个数据库 ： 教师插入结束");
+                // System.out.println(masterUrl);
+                // System.out.println(dataSourceMap.get(1));
+                // System.out.println(dataSource);
+                // insert 信息同步到其他数据库
+                TeacherSyncUtil.TeacherSyncUtilInsert(teachers, dataSourceMap, masterIp, masterUrl);
+                return result;
+            }else {
+                TeacherSyncUtil.TeacherSyncUtilInsert(teachers, dataSourceMap, masterIp, masterUrl);
+                return 1;
+            }
+
         }catch (Exception e){
             // e.printStackTrace();
             return 0;
@@ -61,11 +99,27 @@ public class TeachersService {
 
     @Transactional
     public int deleteTeachers(String teacherId) {
-        int result = teachersMapper.deleteByPrimaryKey(teacherId);
-        // delete 信息同步到其他数据库
-        TeacherSyncUtil.TeacherSyncUtilDelete(teacherId, dataSourceMap, masterIp);
+        try {
+            String masterUrl = dataSource.getConnection().getMetaData().getURL();
+            masterUrl = JdbcUtil.extractHostFromJdbcUrl(masterUrl);
+            if (masterUrl.equals(masterIp)) {
+                int result = teachersMapper.deleteByPrimaryKey(teacherId);
 
-        return result;
+                System.err.println("第一个数据库 ： 教师删除结束" + "      " + result);
+                // delete 信息同步到其他数据库
+                System.err.println(masterUrl + "    删除数据库的masterUrl");
+                TeacherSyncUtil.TeacherSyncUtilDelete(teacherId, dataSourceMap, masterIp, masterUrl);
+
+                return result;
+            }else {
+                TeacherSyncUtil.TeacherSyncUtilDelete(teacherId, dataSourceMap, masterIp, masterUrl);
+                return 1;
+            }
+
+        }catch (Exception e){
+            return 0;
+        }
+
     }
 
 }
